project: sweep the remaining deprecation warnings
Add SDKCompat.h: guarded old->new constant aliases (UITextAttribute*,
NSGregorianCalendar, kCLAuthorizationStatusAuthorized) that keep the original
2014-era names in the source, plus RB_DEPRECATED_BEGIN/END pragma macros for
APIs with no in-place replacement.

- No-replacement APIs wrapped in RB_DEPRECATED_BEGIN/END (the reconstruction
  keeps them as-is): UIWebView (CustomWebView, RecommendWebView,
  RewardNetworkWebViewController, RecommendWebViewController), NSURLConnection /
  +sendSynchronousRequest (HttpConn, ImageDownloader, Downloader,
  AcViewerOptionViewController, RewardNetworkURLConnection, RewardNetworkWebAPI,
  AppDelegate), remote-notification registration (AppDelegate, DownloadMain),
  UIPasteboard setPersistent: (RecommendAdId, RewardNetworkPasteBoard) and the
  deprecated rotation callbacks (RecommendCore, RewardNetwork).
- Modern replacements kept under SDK guards with the original under #else:
  percent-encoding (ConversionView), NSKeyedArchiver/Unarchiver secure-coding
  (RecommendAdId, RewardNetwork, RewardNetworkPasteBoard),
  NSManagedObjectContext initWithConcurrencyType: (AppDelegate), NSURL via
  NSURLComponents (StoreUtil, DevDataDownloader), CADisplayLink
  preferredFramesPerSecond (MainViewController) and UITabBarItem image/
  selectedImage (FriendScoreMainView, Store* controllers).
- Constant renames via SDKCompat aliases (StoreViewController, UnZipArchive,
  SearchView) and the UnZipArchive uLong -Wformat fix.

+++ 3rdparty/ziparchive/UnZipArchive.mm
@@ -11,6 +11,8 @@
 #import <zconf.h>
 #import <zlib.h>
 
+#import "SDKCompat.h"
+
 @interface UnZipArchive (Private)
 
 - (void)OutputErrorMessage:(NSString *)msg;
@@ -180,7 +182,7 @@
     if (_unzFile) {
         unz_global_info globalInfo = {0};
         if (unzGetGlobalInfo(_unzFile, &globalInfo) == UNZ_OK) {
-            NSLog(@"%d entries in the zip file", globalInfo.number_entry);
+            NSLog(@"%lu entries in the zip file", (unsigned long)globalInfo.number_entry);
         }
     }
     return _unzFile != NULL;
